+++ src/main/java/com/example/web/DeleteCategory.java
@@ -16,27 +16,33 @@ public class DeleteCategory extends HttpServlet {
 
     public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
 
-        String cat_id=request.getParameter("cat_id");
+        HttpSession session = request.getSession();
+        if(session!=null && session.getAttribute("user_id")!=null) {
 
-        try{
-            Connection con = jdbc.getConnexion();
+            String cat_id = request.getParameter("cat_id");
 
-            String query2="DELETE FROM articles WHERE cat_id=?";
-            PreparedStatement ps2 = con.prepareStatement(query2);
-            ps2.setString(1, cat_id);
-            int recordNum2 = ps2.executeUpdate();
+            try {
+                Connection con = jdbc.getConnexion();
 
-            String query="DELETE FROM categories WHERE cat_id=?";
-            PreparedStatement ps = con.prepareStatement(query);
-            ps.setString(1, cat_id);
-            int recordNum = ps.executeUpdate();
-            if(recordNum>0){
-                response.sendRedirect("listCategory");
+                String query2 = "DELETE FROM articles WHERE cat_id=?";
+                PreparedStatement ps2 = con.prepareStatement(query2);
+                ps2.setString(1, cat_id);
+                int recordNum2 = ps2.executeUpdate();
+
+                String query = "DELETE FROM categories WHERE cat_id=?";
+                PreparedStatement ps = con.prepareStatement(query);
+                ps.setString(1, cat_id);
+                int recordNum = ps.executeUpdate();
+                if (recordNum > 0) {
+                    response.sendRedirect("listCategory");
+                }
+
+
+            } catch (Exception e) {
+                e.printStackTrace();
             }
-
-
-        }catch (Exception e){
-            e.printStackTrace();
+        }else {
+            response.sendRedirect("ShowArticle");
         }
     }
